+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fixed NablaTheme.Colors being immutable.
+
 
 ## [1.0.0-alpha23] - 2022-11-16